# perf: add string flattening to reduce retained memory (node only)