--- a/src/parser/flatten/browser.ts
+++ b/src/parser/flatten/browser.ts
@@ -1,0 +1,3 @@
+export function flatten (str: string) {
+  return str
+}