@harttle 2019-10-10
:tada: This PR is included in version 9.1.1 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/liquidjs)
- [GitHub release](https://github.com/harttle/liquidjs/releases/tag/v9.1.1)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: